requirements.txt
flask==3.0.0
python-docx==1.0.1
docx2pdf==0.1.8
python-dotenv==1.0.0
gunicorn==21.2.0
